:art: (test) restructure test cases into units and integrations
